fix: add security check to block cross jwch login
