+++ pkg/ai/jwch_tool.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	service "github.com/ACaiCat/tiktok-go/biz/service/user"
+	"github.com/ACaiCat/tiktok-go/pkg/errno"
 )
 
 type JwchLoginInput struct {
@@ -18,8 +19,15 @@ type JwchLoginOutput struct {
 func jwchLoginTool() LocalTool[JwchLoginInput, JwchLoginOutput] {
 	return LocalTool[JwchLoginInput, JwchLoginOutput]{
 		Name:        "internal_jwch_login",
-		Description: "Logs in to the Jwch system using a user ID. Use this as the required login tool when cookies or credentials are unavailable in the context.",
-		Func:        JwchLogin,
+		Description: "Logs in to the Jwch system using one of the two user IDs in the current chat only. Use this as the required login tool when cookies or credentials are unavailable in the context.",
+		Authorize: func(callCtx ToolCallContext, input JwchLoginInput) error {
+			if !callCtx.CanAccessUser(input.UserID) {
+				return errno.AuthErr.WithMessage("AI只能登录当前对话双方的教务处账号")
+			}
+
+			return nil
+		},
+		Func: JwchLogin,
 	}
 }
