+++ pkg/constants/ai.go
@@ -13,7 +13,7 @@ const (
 教务处工具：
 
 1. 如果你的上下文没有jwch_cookie和jwch_id的信息，那么你就需要调用教务处登录
-2. 你只能使用用户ID登录教务处
+2. 你只能使用当前对话双方的用户ID登录教务处
 3. 每次会话不保存jwch_cookie和jwch_id的信息，你需要重新登录
 
 回复要求：
